remove easy adapter and start using ultimate easyadpater
