+++ AdvancedSlidingMenu/gradle.properties
@@ -1,6 +1,6 @@
 # Project-wide Gradle settings.
-VERSION_NAME=2.0.1
-VERSION_CODE=51
+VERSION_NAME=2.0.2
+VERSION_CODE=52
 GROUP=com.hkm.slidingmenulib
 PUBLISH_GROUP_ID=com.hkm.slidingmenulib
 POM_DESCRIPTION=This is the slider for JAVA android use. Thanks to Android Image Slider. This will make it even a better one!
